Move the S3 build cache provider into its own folder.
