+++ apps/rush-lib/src/api/BuildCacheConfiguration.ts
@@ -26,11 +26,11 @@ import type {
   AzureEnvironmentNames,
   AzureStorageBuildCacheProvider
 } from '../logic/buildCache/AzureStorageBuildCacheProvider';
-const AmazonS3BuildCacheProviderModule: typeof import('../logic/buildCache/AmazonS3BuildCacheProvider') = Import.lazy(
-  '../logic/buildCache/AmazonS3BuildCacheProvider',
+const AmazonS3BuildCacheProviderModule: typeof import('../logic/buildCache/AmazonS3/AmazonS3BuildCacheProvider') = Import.lazy(
+  '../logic/buildCache/AmazonS3/AmazonS3BuildCacheProvider',
   require
 );
-import type { AmazonS3BuildCacheProvider } from '../logic/buildCache/AmazonS3BuildCacheProvider';
+import type { AmazonS3BuildCacheProvider } from '../logic/buildCache/AmazonS3/AmazonS3BuildCacheProvider';
 
 /**
  * Describes the file structure for the "common/config/rush/build-cache.json" config file.
