+++ apps/rush-lib/src/logic/buildCache/AmazonS3/test/AmazonS3BuildCacheProvider.test.ts
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved. Licensed under the MIT license.
 // See LICENSE in the project root for license information.
 
-import { EnvironmentConfiguration } from '../../../api/EnvironmentConfiguration';
+import { EnvironmentConfiguration } from '../../../../api/EnvironmentConfiguration';
 import { AmazonS3BuildCacheProvider } from '../AmazonS3BuildCacheProvider';
 import { StringBufferTerminalProvider, Terminal } from '@rushstack/node-core-library';
-import { RushUserConfiguration } from '../../../api/RushUserConfiguration';
-import { CredentialCache } from '../../CredentialCache';
+import { RushUserConfiguration } from '../../../../api/RushUserConfiguration';
+import { CredentialCache } from '../../../CredentialCache';
 
 describe('AmazonS3BuildCacheProvider', () => {
   let buildCacheWriteCredentialEnvValue: string | undefined;
